Refactor: Replace redundant error mappings, update iter handling, and improve type safety
- Simplified repetitive `.map_err(|e| Error::X(e))` patterns using direct `Error::X`.
- Replaced manual iteration bounds with `.iter()` and `iter().take(..).flatten()` for clarity.
- Updated struct and enum defaults using `#[derive(Default)]` and `#[default]` attributes.
- Boxed large structs (`Box<LocalSandbox>` etc.) to optimize memory usage.
- Applied additional Clippy allowances (`#[allow(clippy::too_many_arguments, clippy::type_complexity)]`) for select modules.
- Migrated to new iterator patterns and conforming ranges (`a..b` replaced with `.contains(&x)` logic).

+++ src/network/mod.rs
@@ -99,11 +99,7 @@ impl TapDevice {
 
         unsafe {
             // Copy name into the ifreq name field
-            libc::strncpy(
-                ifr.ifr_name.as_mut_ptr(),
-                cname.as_ptr(),
-                libc::IFNAMSIZ as usize,
-            );
+            libc::strncpy(ifr.ifr_name.as_mut_ptr(), cname.as_ptr(), libc::IFNAMSIZ);
 
             // Set flags to create a TAP device without extra packet info header.
             ifr.ifr_ifru.ifru_flags = (libc::IFF_TAP | libc::IFF_NO_PI) as libc::c_short;
